Replace primary key type with long for all tables

+++ src/main/resources/db/migration/V1__Create_Tables.sql
@@ -1,6 +1,6 @@
 create table user
 (
-    `id`                      int                 not null primary key auto_increment,
+    `id`                      bigint              not null primary key auto_increment,
     `username`                varchar(50) unique  not null,
     `password`                char(60)            not null,              -- BCrypt password field
     `email`                   varchar(254) unique not null,
@@ -14,11 +14,11 @@ create table user
 
 create table post
 (
-    `id`        int          not null primary key auto_increment,
+    `id`        bigint       not null primary key auto_increment,
     `title`     varchar(100) not null,
     `imageId`   varchar(36), -- Refers to an image named with a UUID, without the file extension
     `text`      text,
-    `user`      int          not null,
+    `user`      bigint       not null,
     `timestamp` datetime     not null,
     `deleted`   boolean      not null default false,
     foreign key (user) references user (id)
@@ -26,11 +26,11 @@ create table post
 
 create table comment
 (
-    `id`        int           not null primary key auto_increment,
+    `id`        bigint        not null primary key auto_increment,
     `imageId`   varchar(36),
     `text`      varchar(5000) not null,
-    `user`      int           not null,
-    `post`      int           not null,
+    `user`      bigint        not null,
+    `post`      bigint        not null,
     `timestamp` datetime      not null,
     `deleted`   boolean       not null default false,
     foreign key (user) references user (id),
@@ -40,10 +40,10 @@ create table comment
 -- Likes purchased by the user
 create table purchase
 (
-    `id`          int            not null primary key auto_increment,
+    `id`          bigint         not null primary key auto_increment,
     `pricePaid`   decimal(13, 4) not null, -- GAAP compatible money field
     `likesBought` int            not null,
-    `user`        int            not null,
+    `user`        bigint         not null,
     `timestamp`   datetime       not null,
     foreign key (user) references user (id)
 );
@@ -51,10 +51,10 @@ create table purchase
 -- Likes spent on post
 create table postlikes
 (
-    `id`        int not null primary key auto_increment,
-    `likesUsed` int not null,
-    `post`      int not null,
-    `user`      int not null,
+    `id`        bigint not null primary key auto_increment,
+    `likesUsed` int    not null,
+    `post`      bigint not null,
+    `user`      bigint not null,
     unique (post, user),
     foreign key (post) references post (id),
     foreign key (user) references user (id)
